Moving the updatestatus and kafka publishing routines into a common abstract test runner
