+++ galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/TestRunner.java
@@ -172,6 +172,7 @@ public class TestRunner extends AbstractTestRunner {
 
             logger.debug("Getting test annotations..");
             IAnnotationExtractor annotationExtractor = dataProvider.getAnnotationExtractor();
+            
             Test testAnnotation = annotationExtractor.getAnnotation( testClass , Test.class);
             logger.debug("Test annotations.. got");
 
@@ -525,101 +526,6 @@ public class TestRunner extends AbstractTestRunner {
         }
     }
 
-    private void markWaiting(@NotNull IFramework framework) throws TestRunException {
-        int initialDelay = 600;
-        int randomDelay = 180;
-
-        DssUtils.incrementMetric(dss, "metrics.runs.made.to.wait");
-
-        try {
-            String sInitialDelay = AbstractManager.nulled(this.cps.getProperty("waiting.initial", "delay"));
-            String sRandomDelay = AbstractManager.nulled(this.cps.getProperty("waiting.random", "delay"));
-
-            if (sInitialDelay != null) {
-                initialDelay = Integer.parseInt(sInitialDelay);
-            }
-            if (sRandomDelay != null) {
-                randomDelay = Integer.parseInt(sRandomDelay);
-            }
-        } catch (Exception e) {
-            logger.error("Problem reading delay properties", e);
-        }
-
-        int totalDelay = initialDelay + framework.getRandom().nextInt(randomDelay);
-        logger.info("Placing this run on waiting for " + totalDelay + " seconds");
-
-        Instant until = Instant.now();
-        until = until.plus(totalDelay, ChronoUnit.SECONDS);
-
-        HashMap<String, String> properties = new HashMap<>();
-        properties.put("run." + run.getName() + ".status", "waiting");
-        properties.put("run." + run.getName() + ".wait.until", until.toString());
-        try {
-            this.dss.put(properties);
-        } catch (DynamicStatusStoreException e) {
-            throw new TestRunException("Unable to place run in waiting state", e);
-        }
-    }
-
-    private void updateStatus(TestRunLifecycleStatus status, String timestamp) throws TestRunException {
-
-        this.testStructure.setStatus(status.toString());
-        if ("finished".equals(status.toString())) {
-            updateResult();
-            this.testStructure.setEndTime(Instant.now());
-        }
-
-        writeTestStructure();
-
-        try {
-            this.dss.put("run." + run.getName() + ".status", status.toString());
-            if (timestamp != null) {
-                this.dss.put("run." + run.getName() + "." + timestamp, Instant.now().toString());
-            }
-        } catch (DynamicStatusStoreException e) {
-            throw new TestRunException("Failed to update status", e);
-        }
-
-        try {
-            produceTestRunLifecycleStatusChangedEvent(status);
-        } catch (TestRunException e) {
-            logger.error("Unable to produce a test run lifecycle status changed event to the Events Service", e);
-        }
-    }
-
-    private void produceTestRunLifecycleStatusChangedEvent(TestRunLifecycleStatus status) throws TestRunException {
-        if (this.isProduceEventsEnabled) {
-            logger.debug("Producing a test run lifecycle status change event.");
-
-            String message = String.format("Galasa test run %s is now in status: %s.", framework.getTestRunName(), status.toString());
-            TestRunLifecycleStatusChangedEvent testRunLifecycleStatusChangedEvent = new TestRunLifecycleStatusChangedEvent(this.cps, Instant.now().toString(), message);
-            String topic = testRunLifecycleStatusChangedEvent.getTopic();
-
-            if (topic != null) {
-                try {
-                    framework.getEventsService().produceEvent(topic, testRunLifecycleStatusChangedEvent);
-                } catch (EventsException e) {
-                    throw new TestRunException("Failed to publish a test run lifecycle status changed event to the Events Service", e);
-                }
-            }
-        }
-    }
-
-    private void updateResult() throws TestRunException {
-        try {
-            if (this.testStructure.getResult() == null) {
-                this.testStructure.setResult("UNKNOWN");
-            }
-            this.dss.put("run." + run.getName() + ".result", this.testStructure.getResult());
-        } catch (DynamicStatusStoreException e) {
-            throw new TestRunException("Failed to update result", e);
-        }
-    }
-
-
-
-
-
     /**
      * Get the test class from the supplied bundle
      * 
@@ -657,12 +563,5 @@ public class TestRunner extends AbstractTestRunner {
         this.bundleContext = context;
     }
 
-    protected IFramework getFramework() {
-        return this.framework;
-    }
-
-    public IConfigurationPropertyStoreService getCPS() {
-        return this.cps;
-    }
     
 }
